EntityBatcher

+++ pwiz_tools/SkylineApi/SkydbApi/Orm/Entity.cs
@@ -4,7 +4,7 @@ namespace SkydbApi.Orm
 {
     public class Entity
     {
-        [Id(TypeType = typeof(long), Column = "Id", Name = "Id")]
+        [Id(TypeType = typeof(int), Column = "Id", Name = "Id")]
         [Generator(Class = "identity")]
         public virtual long? Id { get; set; }
     }
